requirements.txt
beautifulsoup4==4.11.1
bitstring==4.0.0
bs4==0.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
ecdsa==0.18.0
esptool==4.3
Flask==2.2.2
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
numpy==1.24.0
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
pdf2image==1.16.0
pdfminer==20191125
Pillow==9.3.0
poppler-utils==0.1.0
pycparser==2.21
pycryptodome==3.16.0
PyPDF2==3.0.0
pyserial==3.5
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7
reedsolo==1.5.4
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.3.2.post1
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
